requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20

# Web Scraping & Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdownify>=0.11.6

# Geolocation
geopy>=2.4.0

# Utilities
tqdm>=4.66.0
typing-extensions>=4.9.0
